requirements.txt
flask
gunicorn
redis
prometheus_client

python-gnupg